+++ wpimath/src/main/native/cpp/estimator/DifferentialDrivePoseEstimator.cpp
@@ -7,9 +7,10 @@
 
 #include "frc/estimator/DifferentialDrivePoseEstimator.h"
 
+#include <wpi/timestamp.h>
+
 #include "frc/StateSpaceUtil.h"
 #include "frc/estimator/AngleStatistics.h"
-#include "frc2/Timer.h"
 
 using namespace frc;
 
@@ -73,7 +74,7 @@ Pose2d DifferentialDrivePoseEstimator::Update(
     const Rotation2d& gyroAngle,
     const DifferentialDriveWheelSpeeds& wheelSpeeds,
     units::meter_t leftDistance, units::meter_t rightDistance) {
-  return UpdateWithTime(frc2::Timer::GetFPGATimestamp(), gyroAngle, wheelSpeeds,
+  return UpdateWithTime(units::microsecond_t(wpi::Now()), gyroAngle, wheelSpeeds,
                         leftDistance, rightDistance);
 }
