+++ wpimath/src/main/native/include/frc/estimator/MecanumDrivePoseEstimator.h
@@ -17,7 +17,6 @@
 #include "frc/geometry/Pose2d.h"
 #include "frc/geometry/Rotation2d.h"
 #include "frc/kinematics/MecanumDriveKinematics.h"
-#include "frc2/Timer.h"
 
 namespace frc {
